Rename module to langstream
