+++ runtime/runtime-impl/pom.xml
@@ -246,7 +246,7 @@
               <target>
                 <echo>copy python runtime</echo>
                 <mkdir dir="${basedir}/target/python" />
-                <copydir src="${basedir}/src/main/python/sga_runtime" dest="${basedir}/target/python/sga_runtime" />
+                <copydir src="${basedir}/src/main/python/langstream" dest="${basedir}/target/python/langstream" />
                 <copydir src="${basedir}/src/main/python/tests" dest="${basedir}/target/python/tests" />
                 <copy file="${basedir}/src/main/python/tox.ini" todir="${basedir}/target/python" />
                 <copy file="${basedir}/src/main/python/Pipfile" todir="${basedir}/target/python" />
