+++ runtime/runtime-impl/src/main/java/com/datastax/oss/sga/runtime/agent/AgentRunner.java
@@ -206,7 +206,7 @@ public class AgentRunner
         // copy input/output to standard input/output of the java process
         // this allows to use "kubectl logs" easily
         ProcessBuilder processBuilder = new ProcessBuilder(
-                "python3", "-m", "sga_runtime", podRuntimeConfiguration.toAbsolutePath().toString())
+                "python3", "-m", "langstream", podRuntimeConfiguration.toAbsolutePath().toString())
                 .inheritIO()
                 .redirectOutput(ProcessBuilder.Redirect.INHERIT)
                 .redirectError(ProcessBuilder.Redirect.INHERIT);
