+++ runtime/runtime-impl/src/main/python/langstream/internal/kafka_serialization.py
