+++ runtime/runtime-impl/src/main/python/langstream/internal/source_record_tracker.py
@@ -17,7 +17,7 @@
 
 from typing import List, Tuple
 
-from .api import CommitCallback, Record, Source
+from ..api import CommitCallback, Record, Source
 
 
 class SourceRecordTracker(CommitCallback):
