+++ runtime/runtime-impl/src/main/python/langstream/util.py
