+++ runtime/runtime-impl/src/main/python/tests/test_source_record_tracker.py
@@ -17,9 +17,9 @@
 
 from typing import List
 
-from sga_runtime.api import Source, Record
-from sga_runtime.util import SimpleRecord
-from sga_runtime.source_record_tracker import SourceRecordTracker
+from langstream.api import Source, Record
+from langstream.util import SimpleRecord
+from langstream.internal.source_record_tracker import SourceRecordTracker
 
 
 def test_tracker():
